pyproject.toml
[project]
name = "deep-research-ai-agent"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.10"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "google-generativeai>=0.8.5",
    "langchain>=0.1.0",
    "langchain-community>=0.3.27",
    "langchain-openai>=0.3.30",
    "langgraph>=0.0.40",
    "numpy>=2.2.6",
    "openai>=1.100.0",
    "pypdf>=6.0.0",
    "python-docx>=1.2.0",
    "python-dotenv>=1.1.1",
    "tavily-python>=0.3.0",
    "weaviate-client>=3.26.7,<4.0.0",
]
